payments: Move ReimbursementRule class to `models` module
This is to avoid import loops.
